Moved Type tests from tables.rs to new types.rs
